fix : 카테고리 아이콘 파일명 변경 반영 및 presets 20→15개 축소 #56
- Planet 1~20.svg 삭제 → planet-01~15.svg로 교체
- _assetNames 매핑을 실제 파일명과 일치하도록 수정
- presets 목록을 에셋 수(15개)에 맞춰 축소

+++ lib/features/todo/presentation/widgets/category_add_bottom_sheet.dart
@@ -116,7 +116,7 @@ class _CategoryAddBottomSheetState extends State<CategoryAddBottomSheet> {
                   ),
                 ),
               ),
-              SizedBox(height: AppSpacing.s8),
+              SizedBox(height: AppSpacing.s16),
               SizedBox(
                 height: 48.h,
                 child: ListView.builder(
